pyproject.toml
[tool.poetry]
name = "test-project-for-company"
version = "0.1.0"
description = ""
authors = ["ImperatorNeron <katebarspk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
scons = "^4.8.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
